1.added isverified to user model 2.added get school by id api 3.made changes to get users with same school 4.'Send request approval to existing user' api is added 5.Make approval or decline for image user has uploaded is added

+++ documentVerification/document.verification.service.js
@@ -4,44 +4,67 @@ const Hobby = db.Hobby
 const User = db.User
 const fs = require('fs')
 module.exports = {
-    docUpload
+    docUpload,
+    docAction
 };
 
 async function docUpload(req){
     console.log("req.body=>",req.body)
     const documentVerification = new DocumentVerification(req.body);
-    
-    var docData = fs.readFileSync('assets/Images/'+req.file.filename);
-    var document = {
-        data:docData,
-        contentType:req.file.mimetype
+    if(req.file.filename){
+        var docData = fs.readFileSync('assets/Images/'+req.file.filename);
+        var document = {
+            data:docData,
+            contentType:req.file.mimetype
+        }
+        Object.assign(documentVerification,{document:document,status:"Pending"});
+        documentVerification.save();
+        // console.log("req.body => ",req.body)
+        // console.log("req.file => ",req.file)
+        return new Promise((resolve, reject) => {
+            fs.unlink('assets/Images/'+req.file.filename, (err) => {
+                if (err){
+                    // throw err;
+                    reject({
+                        status:"fail",
+                        message :"uploaded the file successfully,but error while deleting redundant file," + err
+                    })
+                } else{
+                    console.log('assets/Images/'+req.file.filename+' was deleted');
+                    resolve({
+                        success:"success",
+                        message :"successfully uploaded document for verification"
+                    });
+                }
+            });
+        })
+    }else{
+        return{
+            status:"fail",
+            message:"Unable to extract the file attached,please try again"
+        }
     }
-    Object.assign(documentVerification,{document:document,status:"Pending"});
-    documentVerification.save();
-    // console.log("req.body => ",req.body)
-    // console.log("req.file => ",req.file)
-    return new Promise((resolve, reject) => {
-        fs.unlink('assets/Images/'+req.file.filename, (err) => {
-            if (err){
-                // throw err;
-                reject({
-                    status:"fail",
-                    message :"uploaded the file successfully,but error while deleting redundant file," + err
-                })
-            } else{
-                console.log('assets/Images/'+req.file.filename+' was deleted');
-                resolve({
-                    success:"success",
-                    message :"successfully uploaded document for verification"
-                });
-            }
-          });
-      })
-    
-    
-    
 
+}
 
+async function docAction({userId,school,status}){
+    const documentVerification = await DocumentVerification.findOne({
+        id:userId,
+        school:school,
+    })
+    if(documentVerification){
+        Object.assign(documentVerification,{status:status})
+        await documentVerification.save()
+        return{
+            status:"success",
+            message:"successfully updated the status to "+ status
+        }
+    }else{
+        return{
+            status:"fail",
+            message:"unable to find any document for the given user and school "+userId
+        }
+    }
 }
